pyproject.toml
[build-system]
requires = ["setuptools>=61.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.distutils.bdist_wheel]
universal = true

[project]
name = "django-vendor-promo"
version = "0.2.0"

authors = [
  { name="Grant Viklund", email="renderbox@gmail.com" },
  { name="Roberto Himmelbauer" }
]
description = "Extension to Django Vendor to add Promo Code capabilities"
readme = "README.md"
requires-python = ">=3.7"
classifiers = [
    "Programming Language :: Python :: 3 :: Only",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
keywords = ["django", "app"]
dependencies = [
    "Django>=3.0,<4.0",
    "django-autoslug",
    "django-extensions",
    "django-vendor>=0.2.25",
    "django-site-configs",
    'django-integrations',
    "iso4217",
    "authorizenet",
    "stripe"
]

[project.optional-dependencies]
dev = [
    "black",
    "crispy-bootstrap4",
    "crispy-bootstrap5",
    "flake8",
    "flake8-black",
    "mypy",
    "bandit",
    "isort",
    "dj-database-url",
    "psycopg2-binary",
    "django-crispy-forms",
    "django-site-configs",
    "django-integrations",
    "django-allauth",
    "toml",
]
docs= [
    "recommonmark",
    "m2r",
    "django_extensions",
    "coverage",
    "Sphinx",
    "rstcheck",
    "sphinx-rtd-theme",
]

[project.urls]
"Homepage" = "https://github.com/renderbox/django-vendor-promo/"
"Bug Tracker" = "https://github.com/renderbox/django-vendor-promo/issues"

[tool.setuptools.package-data]
vendorpromo = ["templates/vendorpromo/*.html", "templates/vendorpromo/*/*.html"]
